+++ crates/nostr-sdk/examples/aggregated-query.rs
@@ -11,7 +11,7 @@ use nostr_sdk::prelude::*;
 async fn main() -> Result<()> {
     tracing_subscriber::fmt::init();
 
-    let database = NostrLMDB::open("./db/nostr-lmdb")?;
+    let database = NostrLMDB::open("./db/nostr-lmdb").await?;
     let client: Client = Client::builder().database(database).build();
     client.add_relay("wss://relay.damus.io").await?;