+++ crates/nostr-sdk/examples/lmdb.rs
@@ -11,7 +11,7 @@ async fn main() -> Result<()> {
 
     let keys = Keys::parse("nsec1ufnus6pju578ste3v90xd5m2decpuzpql2295m3sknqcjzyys9ls0qlc85")?;
 
-    let database = NostrLMDB::open("./db/nostr-lmdb")?;
+    let database = NostrLMDB::open("./db/nostr-lmdb").await?;
     let client: Client = ClientBuilder::default()
         .signer(keys.clone())
         .database(database)